requirements.txt
annotated-types==0.5.0
anyio==3.7.1
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.103.1
h11==0.14.0
httptools==0.6.0
idna==3.4
mypy==1.5.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pydantic==2.3.0
pydantic_core==2.6.3
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
types-requests==2.31.0.2
types-urllib3==1.26.25.14
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
